refactor addBook function to a module, import add_book.js module to index.js
